+++ doc/release_notes.rst
@@ -9,6 +9,8 @@ Release Notes
 .. Upcoming Release
 .. =================
 
+- Added solar rooftop ratio setting to ``add_existing_baseyear`` for heuristically splitting existing solar capacity between rooftop and utility-scale (defaults to a 50:50 split).
+
 PyPSA-Eur v2026.02.0 (18th February 2026)
 =========================================